Update AllianceV2(No blocking, just farming)
Please reduce the code size on this without removing any functionality unless it is deemed unnecessary or insignificant. The goal is to reduce the code size of this code down to 80 percent. Then I can further optimize it's "square chooser" for loops at the beginning to select better target squares for the satellite to drill it.